+++ packages/tswebextension/src/lib/mv2/background/web-request-api.ts
@@ -359,7 +359,7 @@ export class WebRequestApi {
         // the response in order to actually apply $replace rules to it.
         const response = RequestBlockingApi.getBlockingResponse({
             rule: basicResult,
-            popupRule: result.popupRule,
+            popupRule: result.getPopupRule(),
             eventId,
             requestUrl,
             referrerUrl,